Merge from Upstream

+++ .build/psake.ps1
@@ -24,7 +24,7 @@ Task BuildDebianPackage -Depends Test {
 
     # Build debian package structure
     $null = New-Item -ItemType Directory -Path ./deb/automatedlab/usr/local/share/powershell/Modules -Force
-    $null = New-Item -ItemType Directory -Path ./deb/automatedlab/usr/share/AutomatedLab/assets -Force
+    $null = New-Item -ItemType Directory -Path ./deb/automatedlab/usr/share/AutomatedLab/Assets -Force
     $null = New-Item -ItemType Directory -Path ./deb/automatedlab/usr/share/AutomatedLab/Stores -Force
     $null = New-Item -ItemType Directory -Path ./deb/automatedlab/usr/share/AutomatedLab/Labs -Force
     $null = New-Item -ItemType Directory -Path ./deb/automatedlab/usr/share/AutomatedLab/LabSources -Force
@@ -59,7 +59,7 @@ Installed-Size: $('{0:0}' -f ((Get-ChildItem -Path $env:APPVEYOR_BUILD_FOLDER -E
     $confPath = "./deb/automatedlab/usr/local/share/powershell/Modules/AutomatedLab/$($env:APPVEYOR_BUILD_VERSION)/AutomatedLab.init.ps1"
     Add-Content -Path $confPath -Value 'Set-PSFConfig -Module AutomatedLab -Name LabSourcesLocation -Description "Location of lab sources folder" -Validation string -Value "/usr/share/AutomatedLab/LabSources"'
 
-    Copy-Item -Path ./Assets/* -Recurse -Destination ./deb/automatedlab/usr/share/AutomatedLab/assets -Force
+    Copy-Item -Path ./Assets/* -Recurse -Destination ./deb/automatedlab/usr/share/AutomatedLab/Assets -Force
     Copy-Item -Path ./LabSources/* -Recurse -Destination ./deb/automatedlab/usr/share/AutomatedLab/LabSources -Force
 
     # Update permissions on AL folder to allow non-root access to configs
@@ -76,26 +76,30 @@ Task Test -Depends Init {
     "`n`tSTATUS: Testing with PowerShell $PSVersion"
 
     # Ensure recent Pester version is actually used
-    Import-Module -Name Pester -MinimumVersion 4.0.0 -Force
-
-    # Gather test results. Store them in a variable and file
-    $TestResults = Invoke-Pester -Path $ProjectRoot\Tests -PassThru -OutputFormat NUnitXml -OutputFile "$ProjectRoot\$TestFile"
-
-    # In Appveyor?  Upload our tests! #Abstract this into a function?
-    If ($ENV:BHBuildSystem -eq 'AppVeyor')
+    if (-not $IsLinux)
     {
-        (New-Object 'System.Net.WebClient').UploadFile(
-            "https://ci.appveyor.com/api/testresults/nunit/$($env:APPVEYOR_JOB_ID)",
-            "$ProjectRoot\$TestFile" )
-    }
+        Import-Module -Name Pester -MinimumVersion 5.0.0 -Force
 
-    Remove-Item "$ProjectRoot\$TestFile" -Force -ErrorAction SilentlyContinue
+        # Gather test results. Store them in a variable and file
+        $TestResults = Invoke-Pester -Path $ProjectRoot\Tests | ConvertTo-NUnitReport -ErrorAction SilentlyContinue
+        $TestResults.Save("$ProjectRoot\$TestFile")
 
-    # Failed tests?
-    # Need to tell psake or it will proceed to the deployment. Danger!
-    if ($TestResults.FailedCount -gt 0)
-    {
-        throw "Failed '$($TestResults.FailedCount)' tests, build failed"
+        # In Appveyor?  Upload our tests! #Abstract this into a function?
+        If ($ENV:BHBuildSystem -eq 'AppVeyor')
+        {
+            (New-Object 'System.Net.WebClient').UploadFile(
+                "https://ci.appveyor.com/api/testresults/nunit/$($env:APPVEYOR_JOB_ID)",
+                "$ProjectRoot\$TestFile" )
+        }
+
+        Remove-Item "$ProjectRoot\$TestFile" -Force -ErrorAction SilentlyContinue
+
+        # Failed tests?
+        # Need to tell psake or it will proceed to the deployment. Danger!
+        if ($TestResults.FailedCount -gt 0)
+        {
+            throw "Failed '$($TestResults.FailedCount)' tests, build failed"
+        }
     }
     "`n"
 }
